st: share azaghal entity code between rare and rno2
