+++ config/symbols.us.strno2.txt
@@ -16,7 +16,17 @@ g_EInitFlyingZombieHalf2 = 0x80180958;
 g_EInitFlyingZombieHalf1 = 0x80180964;
 g_EInitGhostDancer = 0x80180988;
 BackgroundBlockInit = 0x801809B0;
+sensors_ground = 0x80181FF4;
 body_config = 0x80182004;
+sword_face_tl = 0x80182034;
+sword_face_tr = 0x8018203C;
+sword_face_bl = 0x80182044;
+sword_face_br = 0x8018204C;
+sword_edge_tl = 0x80182054;
+sword_edge_tr = 0x8018205C;
+sword_edge_bl = 0x80182064;
+sword_edge_br = 0x8018206C;
+sword_tip = 0x80182074;
 anim_seek_player = 0x80182140;
 anim_sword_overhead_a = 0x80182168;
 anim_sword_overhead_b = 0x80182170;
@@ -32,6 +42,7 @@ anim_slash_up_b = 0x801821E8;
 anim_fall_backward = 0x80182218;
 anim_combo_init = 0x80182230;
 anim_combo = 0x80182250;
+indices = 0x80182298;
 rooms = 0x80182390;
 EntityBreakable = 0x801B22E4;
 EntityBreakableDebris = 0x801B2664;
